+++ HVLib/Types/Blob/HVBlobPutParameters.h
@@ -23,10 +23,10 @@
 @interface HVBlobHashAlgorithmParameters : HVType
 {
 @private
-    int m_blockSize;
+    NSInteger m_blockSize;
 }
 
-@property (readwrite, nonatomic) int blockSize;
+@property (readwrite, nonatomic) NSInteger blockSize;
 
 @end
 
@@ -34,15 +34,15 @@
 {
 @private
     NSString* m_url;
-    int m_chunkSize;
-    int m_maxSize;
+    NSInteger m_chunkSize;
+    NSInteger m_maxSize;
     NSString* m_hashAlgorithm;
     HVBlobHashAlgorithmParameters* m_hashParams;
 }
 
 @property (readwrite, nonatomic, retain) NSString* url;
-@property (readwrite, nonatomic) int chunkSize;
-@property (readwrite, nonatomic) int maxSize;
+@property (readwrite, nonatomic) NSInteger chunkSize;
+@property (readwrite, nonatomic) NSInteger maxSize;
 @property (readwrite, nonatomic, retain) NSString* hashAlgorithm;
 @property (readwrite, nonatomic, retain) HVBlobHashAlgorithmParameters* hashParams;
